[Refactor] Merge offline PD tests into main test file
Consolidate test_qwen3_omni_pd.py into test_qwen3_omni.py for offline
inference, reducing file count and keeping PD tests alongside existing
tests.
