+++ tests/e2e/offline_inference/test_qwen3_omni.py
@@ -1,5 +1,8 @@
 """
 E2E offline tests for Omni model with video input and audio output.
+
+Includes PD (Prefill-Decode) disaggregation tests that use a separate
+stage config YAML splitting the thinker into prefill + decode stages.
 """
 
 import os
@@ -29,10 +32,15 @@ else:
 # Create parameter combinations for model and stage config
 test_params = [(model, stage_config) for model in models for stage_config in stage_configs]
 
+# PD disaggregation stage config (requires 3x GPUs)
+pd_stage_configs = [str(Path(__file__).parent.parent / "stage_configs" / "qwen3_omni_pd_ci.yaml")]
+pd_test_params = [(model, sc) for model in models for sc in pd_stage_configs]
+
 
 def get_question(prompt_type="video"):
     prompts = {
         "video": "Describe the video briefly.",
+        "text": "What is the capital of China? Answer in 20 words.",
     }
     return prompts.get(prompt_type, prompts["video"])
 
@@ -48,4 +56,52 @@ def test_video_to_audio(omni_runner, omni_runner_handler) -> None:
     request_config = {"prompts": get_question(), "videos": video, "modalities": ["audio"]}
 
     # Test single completion
+    omni_runner_handler.send_request(request_config)
+
+
+# ===================================================================
+# PD (Prefill-Decode) disaggregation tests
+# ===================================================================
+
+
+@pytest.mark.core_model
+@pytest.mark.omni
+@hardware_test(res={"cuda": "H100"}, num_cards=3)
+@pytest.mark.parametrize("omni_runner", pd_test_params, indirect=True)
+def test_pd_text_only(omni_runner, omni_runner_handler) -> None:
+    """Test PD disaggregation with text-only output (no talker/code2wav)."""
+    request_config = {
+        "prompts": get_question("text"),
+        "modalities": ["text"],
+    }
+    omni_runner_handler.send_request(request_config)
+
+
+@pytest.mark.core_model
+@pytest.mark.omni
+@hardware_test(res={"cuda": "H100"}, num_cards=3)
+@pytest.mark.parametrize("omni_runner", pd_test_params, indirect=True)
+def test_pd_text_to_audio(omni_runner, omni_runner_handler) -> None:
+    """Test PD disaggregation with text input and audio output
+    through the full 4-stage pipeline (prefill -> decode -> talker -> code2wav)."""
+    request_config = {
+        "prompts": get_question("text"),
+        "modalities": ["audio"],
+    }
+    omni_runner_handler.send_request(request_config)
+
+
+@pytest.mark.core_model
+@pytest.mark.omni
+@hardware_test(res={"cuda": "H100"}, num_cards=3)
+@pytest.mark.parametrize("omni_runner", pd_test_params, indirect=True)
+def test_pd_video_to_audio(omni_runner, omni_runner_handler) -> None:
+    """Test PD disaggregation with video input and audio output
+    through the full 4-stage pipeline (prefill -> decode -> talker -> code2wav)."""
+    video = generate_synthetic_video(224, 224, 300)["np_array"]
+    request_config = {
+        "prompts": get_question("video"),
+        "videos": video,
+        "modalities": ["audio"],
+    }
     omni_runner_handler.send_request(request_config)
